kernel/vm: Cast temporary to uint32.
It is defined as a uint32:1, which apparently becomes an "int" on
newer GCC versions, which thus triggers a -Werror=format. So,
convert it explicitly in order to prevent the error.

+++ src/system/kernel/vm/VMCache.cpp
@@ -1455,7 +1455,7 @@ VMCache::Dump(bool showPages) const
 	kprintf("  type:         %s\n", vm_cache_type_to_string(type));
 	kprintf("  virtual_base: 0x%" B_PRIx64 "\n", virtual_base);
 	kprintf("  virtual_end:  0x%" B_PRIx64 "\n", virtual_end);
-	kprintf("  temporary:    %" B_PRIu32 "\n", temporary);
+	kprintf("  temporary:    %" B_PRIu32 "\n", uint32(temporary));
 	kprintf("  lock:         %p\n", &fLock);
 #if KDEBUG
 	kprintf("  lock.holder:  %" B_PRId32 "\n", fLock.holder);
